website/requirements.txt
absl-py==0.4.0
alabaster==0.7.10
anaconda-client==1.6.9
anaconda-navigator==1.7.0
anaconda-project==0.8.2
appnope==0.1.0
appscript==1.0.1
asn1crypto==0.24.0
astor==0.7.1
astroid==1.6.1
astropy==2.0.3
attrs==17.4.0
audioread==2.1.6
Augmentor==0.2.0
Babel==2.5.3
backports.shutil-get-terminal-size==1.0.0
bcolz==1.2.1
beautifulsoup4==4.6.0
bitarray==0.8.1
bkcharts==0.2
blaze==0.11.3
bleach==1.5.0
bokeh==0.12.13
boto==2.48.0
Bottleneck==1.2.1
branca==0.3.0
certifi==2018.8.24
cffi==1.11.4
chardet==3.0.4
click==6.7
cloudpickle==0.5.2
clyent==1.2.2
colorama==0.3.9
colorcet==0.9.1
conda==4.4.10
conda-build==3.4.1
conda-verify==2.0.0
contextlib2==0.5.5
cryptography==2.1.4
cycler==0.10.0
cymem==1.31.2
Cython==0.27.3
cytoolz==0.8.2
dask==0.16.1
datashader==0.6.8
datashape==0.5.4
decorator==4.2.1
dill==0.2.7.1
distributed==1.20.2
docutils==0.14
elasticsearch==6.3.1
en-core-web-sm==2.0.0
entrypoints==0.2.3
enum34==1.1.6
et-xmlfile==1.0.1
fastai==0.7.0
fastcache==1.0.2
feather-format==0.4.0
filelock==2.0.13
Flask==1.0.2
Flask-Cors==3.0.3
folium==0.6.0
future==0.16.0
gast==0.2.0
gevent==1.2.2
glob2==0.6
gmpy2==2.0.8
graphviz==0.8.3
greenlet==0.4.12
grpcio==1.14.1
h5py==2.7.1
heapdict==1.0.0
html5lib==0.9999999
idna==2.6
imageio==2.2.0
imagesize==0.7.1
ipykernel==4.8.0
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.1
isort==4.2.15
isoweek==1.3.3
itsdangerous==0.24
jdcal==1.3
jedi==0.11.1
Jinja2==2.10
joblib==0.12.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.2
jupyter-console==5.2.0
jupyter-core==4.4.0
jupyterlab==0.31.5
jupyterlab-launcher==0.10.2
Keras==2.0.5
lazy-object-proxy==1.3.1
librosa==0.6.2
llist==0.6
llvmlite==0.24.0
locket==0.2.0
lxml==4.1.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.1.2
mccabe==0.6.1
mistune==0.8.3
mizani==0.4.6
mpmath==1.0.0
msgpack-numpy==0.4.1
msgpack-python==0.5.1
multipledispatch==0.4.9
murmurhash==0.28.0
navigator-updater==0.1.0
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
nltk==3.2.5
nose==1.3.7
notebook==5.5.0
npm==0.1.1
numba==0.39.0
numexpr==2.6.4
numpy==1.15.1
numpydoc==0.7.0
object-detection==0.1
odo==0.5.1
olefile==0.45.1
opencv-python==3.1.0.5
openpyxl==2.4.10
optional-django==0.1.0
packaging==16.8
palettable==3.1.1
pandas==0.22.0
pandas-summary==0.0.41
pandocfilters==1.4.2
param==1.8.1
parse-rest==0.2.20170114
parso==0.1.1
partd==0.3.8
path.py==10.5
pathlib==1.0.1
pathlib2==2.3.0
patsy==0.5.0
pep8==1.7.1
pexpect==4.3.1
pickleshare==0.7.4
Pillow==5.0.0
pkginfo==1.4.1
plac==0.9.6
plotly==2.5.1
plotnine==0.3.0
pluggy==0.6.0
ply==3.10
preshed==1.0.0
prompt-toolkit==1.0.15
protobuf==3.6.1
psutil==5.4.3
ptyprocess==0.5.2
py==1.5.2
pyarrow==0.9.0.post1
pycodestyle==2.3.1
pycosat==0.6.3
pycparser==2.18
pycrypto==2.6.1
pyct==0.4.5
pycurl==7.43.0.1
pyflakes==1.6.0
Pygments==2.2.0
pylint==1.8.2
pyodbc==4.0.22
pyOpenSSL==17.5.0
pyparsing==2.2.0
PySocks==1.6.7
pytest==3.3.2
python-dateutil==2.6.1
pytz==2017.3
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==17.0.0
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.3.1
regex==2017.4.5
requests==2.18.4
resampy==0.2.1
rope==0.10.7
ruamel-yaml==0.15.35
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.0.0
seaborn==0.8.1
Send2Trash==1.4.2
simplegeneric==0.8.1
simplejson==3.15.0
singledispatch==3.4.0.3
six==1.11.0
sklearn-pandas==1.6.0
snowballstemmer==1.2.1
sortedcollections==0.5.3
sortedcontainers==1.5.9
spacy==2.0.11
Sphinx==1.6.6
sphinxcontrib-websupport==1.0.1
spyder==3.2.6
SQLAlchemy==1.2.1
statsmodels==0.8.0
sympy==1.1.1
tables==3.4.2
tblib==1.3.2
tensorboard==1.9.0
tensorflow==1.9.0
tensorflow-tensorboard==0.4.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
Theano==1.0.2
thinc==6.10.2
toolz==0.9.0
torch==0.3.1
torchtext==0.2.3
torchvision==0.2.1
tornado==4.5.3
tqdm==4.19.8
traitlets==4.3.2
typing==3.6.2
ujson==1.35
unicodecsv==0.14.1
Unidecode==1.0.22
urllib3==1.22
virtualenv==16.0.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.1.0
wrapt==1.10.11
WTForms==2.1
xarray==0.10.9
xlrd==1.1.0
XlsxWriter==1.0.2
xlwings==0.11.5
xlwt==1.2.0
zict==0.1.3
